+++ eframe-core/src/main/java/net/linlan/frame/comm/service/PermissionService.java
@@ -20,13 +20,13 @@ package net.linlan.frame.comm.service;
 import java.util.Objects;
 import java.util.Set;
 
-import net.linlan.frame.comm.constant.AdminType;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
 import net.linlan.commons.core.ObjectUtils;
 import net.linlan.commons.core.StringUtils;
 import net.linlan.frame.FrameAdminUser;
+import net.linlan.frame.comm.constant.AdminType;
 import net.linlan.frame.comm.security.context.PermissionContextHolder;
 import net.linlan.frame.web.SecurityUtils;
 import net.linlan.utils.constant.Constants;